+++ node/clients/intelligent-search-api.ts
@@ -80,7 +80,7 @@ export class IntelligentSearchApi extends ExternalClient {
     return this.http.get(`/banners/${path}`, {params: {...params, query: params.query, locale: this.locale}, metric: 'banners'})
   }
 
-  public async facets(params: FacetsArgs, path: string) {
+  public async facets(params: FacetsArgs, path: string, shippingHeader?: string[]) {
     const {query, leap, searchState} = params
 
     return this.http.get(`/facets/${path}`, {
@@ -91,11 +91,14 @@ export class IntelligentSearchApi extends ExternalClient {
         bgy_leap: leap ? true : undefined,
         ...parseState(searchState),
       },
-      metric: 'facets'
+      metric: 'facets',
+      headers: {
+        'x-vtex-shipping-options': shippingHeader ?? '',
+      },
     })
   }
 
-  public async productSearch(params: SearchResultArgs, path: string) {
+  public async productSearch(params: SearchResultArgs, path: string, shippingHeader?: string[]) {
     const {query, leap, searchState} = params
 
     return this.http.get(`/product_search/${path}`, {
@@ -106,7 +109,10 @@ export class IntelligentSearchApi extends ExternalClient {
         ...parseState(searchState),
         ...params,
       },
-      metric: 'product-search'
+      metric: 'product-search',
+      headers: {
+        'x-vtex-shipping-options': shippingHeader ?? '',
+      },
     })
   }
 }